feature: route for survey added and created a component with button to survey
